+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## UNRELEASED
 
+### FIXED
+
+### ADDED
+
+- Worker health check and fetch analysis logs ([Issue](https://github.com/orgs/amosproj/projects/79/views/2?pane=issue&itemId=112364687&issue=amosproj%7Camos2025ss01-embark-orchestration-framework%7C46))
+
 ## [sprint-11](https://github.com/amosproj/amos2025ss01-embark/releases/tag/sprint-11-release) - 2025-07-02
 
 ### FIXED